removed warnings, and fixed ga makefile
git-svn-id: https://groke.otago.ac.nz/ant/ant@247 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/string_pair.h
@@ -21,11 +21,11 @@ public:
 	size_t length(void) { return string_length; }
 	char *string(void) { return start; }
 	char operator[](long pos) { return start[pos]; }
-	char *strcpy(char *dest) { *(strncpy(dest, start, string_length) + string_length + 1) = '\0'; return dest; }
+        char *strcpy(char *dest) { *(strncpy(dest, start, string_length) + string_length + 1) = '\0'; return dest; }
 	int strcmp(ANT_string_pair *with) { return string_length == with->string_length ? ::memcmp(start, with->start, string_length) : string_length < with->string_length ? -1 : 1; }
 	int true_strcmp(ANT_string_pair *with);
 	int strncmp(ANT_string_pair *with, size_t length);
-	void text_render(void) { printf("%*.*s", string_length, string_length, start); }
+	void text_render(void) { printf("%*.*s", (int)string_length, (int)string_length, start); }
 };
 
 /*
